Fixed bugs exposed by VC++
git-svn-id: http://yate.null.ro/svn/yate/trunk@921 acf43c95-373e-0410-b603-e72c3f656dc1

+++ modules/yiaxchan.cpp
@@ -926,7 +926,7 @@ bool YIAXEngine::userreg(IAXTransaction* tr, bool regrel)
 	// TODO: support number != username
 	data << "/" << tr->username();
 	msg.addParam("data",data);
-	msg.addParam("expires",String(tr->expire()));
+	msg.addParam("expires",String((long)tr->expire()));
     }
     msg.addParam("ip_host",tr->remoteAddr().host());
     msg.addParam("ip_port",String(tr->remoteAddr().port()));
